# api-docs.yaml
openapi: 3.1.0
info:
  title: Ai知识服务后端
  version: "0.1"
servers:
- url: http://localhost:9876
  description: Generated server url
...
paths:
...
  /api/v1/assistants:
    put:
      tags:
      - ai助手管理
      operationId: update_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantUpdateParam"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseAssistantVo"
      security:
      - auth: []
    post:
      tags:
      - ai助手管理
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantCreateParam"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseAssistantVo"
      security:
      - auth: []
...
  /api/v1/chat/conversations/search:
    post:
      tags:
      - 对话管理
      operationId: search_4
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchConversation"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponsePageVoConversationVo"
      security:
      - auth: []
  /api/v1/chat/conversations/messages/search:
    post:
      tags:
      - 对话管理
      operationId: messages
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchMessage"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponsePageVoMessageVo"
      security:
      - auth: []
  /api/v1/chat/conversations/chat_with_assistant:
    post:
      tags:
      - 对话管理
      operationId: newChat
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantChatParam"
        required: true
      responses:
        "400":
          description: BAD Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BizError"
        "200":
          description: OK
          content:
            text/event-stream:
              schema:
                anyOf:
                - $ref: "#/components/schemas/RetrievedContent"
                - $ref: "#/components/schemas/PartialMessage"
                - $ref: "#/components/schemas/Finish"
      security:
      - auth: []
  /api/v1/chat/conversations/chat:
    post:
      tags:
      - 对话管理
      operationId: newChat_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatParam"
        required: true
      responses:
        "400":
          description: BAD Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BizError"
        "200":
          description: OK
          content:
            text/event-stream:
              schema:
                anyOf:
                - $ref: "#/components/schemas/RetrievedContent"
                - $ref: "#/components/schemas/PartialMessage"
                - $ref: "#/components/schemas/Finish"
      security:
      - auth: []
...
  /api/v1/assistants/search:
    post:
      tags:
      - ai助手管理
      operationId: search_6
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchAssistant"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponsePageVoAssistantVo"
      security:
      - auth: []
...
  /api/v1/assistants/{id}/detail:
    get:
      tags:
      - ai助手管理
      operationId: search_7
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseAssistantDetailVo"
      security:
      - auth: []
...
  /api/v1/assistants/{id}:
    delete:
      tags:
      - ai助手管理
      operationId: delete_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEmpty"
      security:
      - auth: []
components:
  schemas:
    JsonNode:
      type: object
...
    Empty:
      type: object
    ResponseEmpty:
      type: object
      properties:
        code:
          type: integer
          format: int32
        msg:
          type: string
        data:
          $ref: "#/components/schemas/Empty"
...
    AssistantUpdateParam:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        llmModel:
          type: string
        systemPrompt:
          type: string
        description:
          type: string
        baseIds:
          type: array
          items:
            type: string
        ext:
          $ref: "#/components/schemas/JsonNode"
    AssistantVo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        llmModel:
          type: string
        systemPrompt:
          type: string
        permit:
          type: integer
          format: int32
        description:
          type: string
        creator:
          type: string
        creatorName:
          type: string
        ext:
          $ref: "#/components/schemas/JsonNode"
        createTime:
          type: integer
          format: int64
        updateTime:
          type: integer
          format: int64
    ResponseAssistantVo:
      type: object
      properties:
        code:
          type: integer
          format: int32
        msg:
          type: string
        data:
          $ref: "#/components/schemas/AssistantVo"
...
    UploadResult:
      type: object
      properties:
        url:
          type: string
        id:
          type: string
        fileName:
          type: string
...
    SearchConversation:
      type: object
      properties:
        key:
          type: string
          description: 按名字检索
        pageNo:
          type: integer
          format: int32
          default: "1"
          description: 从1开始
        pageSize:
          type: integer
          format: int32
          default: "20"
        sort:
          type: string
          enum:
          - CREATED_AT_ASC
          - CREATED_AT_DESC
          - UPDATED_AT_ASC
          - UPDATED_AT_DESC
    ConversationVo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        creator:
          type: string
        creatorName:
          type: string
        createTime:
          type: integer
          format: int64
        updateTime:
          type: integer
          format: int64
    PageVoConversationVo:
      type: object
      properties:
        total:
          type: integer
          format: int64
        pageSize:
          type: integer
          format: int64
        pageNo:
          type: integer
          format: int64
        records:
          type: array
          items:
            $ref: "#/components/schemas/ConversationVo"
    ResponsePageVoConversationVo:
      type: object
      properties:
        code:
          type: integer
          format: int32
        msg:
          type: string
        data:
          $ref: "#/components/schemas/PageVoConversationVo"
    SearchMessage:
      type: object
      properties:
        conversationId:
          type: string
        pageNo:
          type: integer
          format: int32
          default: "1"
          description: 从1开始
        pageSize:
          type: integer
          format: int32
          default: "20"
        sort:
          type: string
          enum:
          - CREATED_AT_ASC
          - CREATED_AT_DESC
          - UPDATED_AT_ASC
          - UPDATED_AT_DESC
    MessageVo:
      type: object
      properties:
        id:
          type: string
        conversationId:
          type: string
        query:
          type: string
        answer:
          type: string
        files:
          $ref: "#/components/schemas/QueryContext"
        createTime:
          type: integer
          format: int64
        updateTime:
          type: integer
          format: int64
    PageVoMessageVo:
      type: object
      properties:
        total:
          type: integer
          format: int64
        pageSize:
          type: integer
          format: int64
        pageNo:
          type: integer
          format: int64
        records:
          type: array
          items:
            $ref: "#/components/schemas/MessageVo"
    QueryContext:
      type: object
      properties:
        resources:
          type: array
          items:
            $ref: "#/components/schemas/RetrievalResource"
        files:
          type: array
          items:
            $ref: "#/components/schemas/UploadResult"
    ResponsePageVoMessageVo:
      type: object
      properties:
        code:
          type: integer
          format: int32
        msg:
          type: string
        data:
          $ref: "#/components/schemas/PageVoMessageVo"
    RetrievalResource:
      type: object
      properties:
        documentId:
          type: string
        fileName:
          type: string
        url:
          type: string
    AssistantChatParam:
      type: object
      properties:
        conversationId:
          type: string
        query:
          type: string
        assistantId:
          type: string
        files:
          type: array
          items:
            $ref: "#/components/schemas/UploadResult"
    BizError:
      type: object
      properties:
        code:
          type: integer
          format: int32
        msg:
          type: string
    RetrievedContent:
      type: object
      properties:
        conversationId:
          type: string
        ctx:
          $ref: "#/components/schemas/QueryContext"
    PartialMessage:
      type: object
      properties:
        conversationId:
          type: string
        text:
          type: string
    ChatResponseMetadata:
      type: object
    Finish:
      type: object
      properties:
        conversationId:
          type: string
        metadata:
          $ref: "#/components/schemas/ChatResponseMetadata"
    ChatParam:
      type: object
      properties:
        conversationId:
          type: string
        query:
          type: string
        llmModel:
          type: string
        baseIds:
          type: array
          items:
            type: string
        files:
          type: array
          items:
            $ref: "#/components/schemas/UploadResult"
...
    AssistantCreateParam:
      type: object
      properties:
        name:
          type: string
        llmModel:
          type: string
        systemPrompt:
          type: string
        permit:
          type: integer
          format: int32
        description:
          type: string
        baseIds:
          type: array
          items:
            type: string
        creator:
          type: string
        ext:
          $ref: "#/components/schemas/JsonNode"
    SearchAssistant:
      type: object
      properties:
        key:
          type: string
          description: 按名字检索
        pageNo:
          type: integer
          format: int32
          default: "1"
          description: 从1开始
        pageSize:
          type: integer
          format: int32
          default: "20"
        sort:
          type: string
          enum:
          - CREATED_AT_ASC
          - CREATED_AT_DESC
          - UPDATED_AT_ASC
          - UPDATED_AT_DESC
    PageVoAssistantVo:
      type: object
      properties:
        total:
          type: integer
          format: int64
        pageSize:
          type: integer
          format: int64
        pageNo:
          type: integer
          format: int64
        records:
          type: array
          items:
            $ref: "#/components/schemas/AssistantVo"
    ResponsePageVoAssistantVo:
      type: object
      properties:
        code:
          type: integer
          format: int32
        msg:
          type: string
        data:
          $ref: "#/components/schemas/PageVoAssistantVo"
...
    AssistantDetailVo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        llmModel:
          type: string
        systemPrompt:
          type: string
        permit:
          type: integer
          format: int32
        description:
          type: string
        creator:
          type: string
        creatorName:
          type: string
        ext:
          $ref: "#/components/schemas/JsonNode"
        createTime:
          type: integer
          format: int64
        updateTime:
          type: integer
          format: int64
        bases:
          type: array
          items:
            $ref: "#/components/schemas/BaseInfo"
    BaseInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    ResponseAssistantDetailVo:
      type: object
      properties:
        code:
          type: integer
          format: int32
        msg:
          type: string
        data:
          $ref: "#/components/schemas/AssistantDetailVo"
  securitySchemes:
    auth:
      type: apiKey
      name: auth
      in: header