# src/main/java/ai/zhidun/app/hub/chat/model/MessageVo.java
package ai.zhidun.app.hub.chat.model;

public record MessageVo(
        String id,
        String conversationId,
        String query,
        String answer,
        QueryContext files,
        long createTime,
        long updateTime) {
}